pyproject.toml
[project]
name = "lmp-reproduced"
version = "0.2.0"
description = "LAMMPS Simulation Framework for Metal Matrix Composites"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
    "python-docx>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lmp_reproduced"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["src"]
